requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
apscheduler==3.10.4
playwright==1.44.0
mlxtend==0.23.1
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
